Fix Handling of Multiple Commits per Notif.
Previously, we were only parsing the first commit of the notification. I made a bad assumption that each commit appears in its own table; it's more complicated than that. Each commit is represented as a pair of rows -- one for the header, and the other with the hunks. So... now the code has leveled up to handle this.