Order Commits in Reverse Order instead of Chron
It turns out that ordering commits by chronological order of author date is incorrect, since the order commits were authored is likely not the order they were committed. It would appear that Notifyr presents commits in the order they were committed, with the latest commit at the top; so, now we import commits according to that order.